Make column definitions in Task db model consistent with schema
Two of the column definitions in the Task db model class
were not consistent with the database schema as specified
in the SQL file included in the portal repository.
Change the definitions here so that they match.
In practice, this means making two columns into
"text" type (i.e., having maximum length 65535).