Add H2 database for use in tests
Add the H2 database JAR for use in unit, integration, and
system tests. On shutdown, it has been observed that H2's
watchdog thread can hang around too long, so that Tomcat
observes it and cleans it up. Avoid that by adding a small
sleep in the shutdown method of the context listener.