Be more careful shutting down the application context
Only attempt to shut down the DBContext if it has in
fact been opened. This may prevent an error during
shutdown, if shutdown happens before the database
has been opened.

For the future: detect if log4j has been started, and
if so, shut it down. (Log4j is used by test frameworks.)