Make JsonTreeTransformProvider.Concept a static nested class
JsonTreeTransformProvider.Concept was declared without
the "static" keyword, and was therefore an inner class.
Add the "static" keyword to the class declaration, to make
it a static nested class instead. See The Java Programming
Language 4e, chapter 5. Instances of the Concept class are
not tied to an instance of the parent class.
The inner class ConceptHandler makes use of the logger, so
that can (for now) remain an inner class.
My guess is that it uses more memory to represent instances
of inner classes, so there is a memory benefit in changing
the declaration of the Concept class.