Vocabs CC-1778 CC-1787: Better handling of cycles and polyhierarchies
Work in progress towards a much better "browse tree".
In the first instance, fix the cycle detection, which,
on detection, threw an exception that wasn't caught, leaving
publication only partially done. Now, don't throw an
exception on cycle detection, but don't give back
a concept_tree file, either.

Fix cycle detection, so that it does not return a false positive
on a polyhierarchy.

How is all this done: we now perform a depth-first search
on the graph of concepts. Cycle detection is done by
noting if we find a back edge during the search.
Polyhierarchy detection is based on finding forward and
cross edges.

If we find a polyhierarchy, for now, we don't return a concept_tree
file either. We can and should do better.