Use Commons Configuration to implement ToolkitProperties
Use Apache Commons Configuration as the back-end
implementation of ToolkitProperties.

This enables support for interpolation of properties using
the ${...} syntax, which we will need later.
(We will use it as part of the change to create an
embedded Tomcat to run Sesame, so that we can specify
the port number once, and reuse it in the definition
of other properties.)

Do a cleanup of uses of instances of "PROPS" throughout
the code; this was probably a naughty way of coding, in
any case.