Fix wrong debugging output in TaskRunner
The TaskRunner outputs a debug message
"Import in progress" before invoking an import provider.
But this message "Import in progress" had been copied/pasted
into the transform and publish methods as well, so that
you saw "Import in progress" before transform and publish.
Edited the log messages to match what is actually about
to be done.