Update Checkstyle to 6.19 to match Eclipse plugin
Update Checkstyle library to release 6.19 so as to match
the latest version of the Eclipse Checkstyle plugin.
It is helpful to do this, as it avoids the potential
issue of thinking that there are no violations based
on what Eclipse is saying, then finding after commit
(and the Bamboo build) that "ant checkstyle" fails.