Make nested classes private and static
The Google "error-prone" tool suggests making the
nested classes that are used in transformers as
handlers of RDF statements static. This seems
to be the right thing to do. At the same time,
I made them all private as well. In the case of
JsonTreeTransformProvider, it was then necessary
to add @SuppressWarnings annotations for the
getters, as they are not obviously used -- but
they are used (using reflection) during JSON extraction.