CC-1799: Support fetching from PoolParty 5.5
Change harvest, backup, and metadata transform
providers to cope with the changes made to data
representation in PoolParty release 5.5.

NB: This change affects the toolkit.properties file.

The properties:
* PoolPartyHarvester.remoteUrl
* PoolPartyHarvester.username
* PoolPartyHarvester.password

have been renamed to:
* PoolParty.remoteUrl
* PoolParty.username
* PoolParty.password

respectively.

And, whereas the value of PoolPartyHarvester.remoteUrl was required to
end with "/PoolParty/api/projects/", PoolParty.remoteUrl should be a
URL ending with "/PoolParty/". (I.e., remove the final component
"api/projects/".)