Force PoolParty backup provider to ask for TriG
The PoolParty backup provider used to ask for data
in the format which is the value of the Toolkit property
PoolPartyHarvester.defaultFormat. But if you changed that
to anything other than TriG or TriX, the backup would
certainly not work. So (at least, for now) specify TriG
explicitly. We could add a separate Toolkit property
later, if we thought that would be helpful.