During PoolParty (meta)data harvest, remove existing files
When fetching PoolParty metadata, remove any existing directory
containing a previous fetch.
When fetching PoolParty (meta)data for an exportModule,
remove any existing fetched data for that exportModule.

The significance of/need for this was
found when changing the Toolkit property
PoolPartyHarvester.defaultFormat: any old files corresponding
to the previous setting were left.

Note: if you change the set of exportModules to fetch,
i.e., by removing one, this isn't "clever" enough to remove
an existing fetch of the removed exportModule. That's
because the harvest directory can (in theory) be a mixture
of data fetched from PoolParty and from elsewhere, so we
can't just remove the entire directory.