Remove final qualifier from parameters of abstract methods
Recent Checkstyle releases now complain about the use
of the final qualifier on formal parameters of abstract
methods. So remove that qualifier where there is
a complaint. (NB: this commit was prompted by my having
updated the Checkstyle plugin in my Eclipse installation.
I'm not yet updating the Checkstyle library in this
repository.)