CC-1847: Have PP nightly backup not overwrite Toolkit log
The nightly PoolParty backup script does not specify a separate
logging configuration, so it uses the Toolkit's own logging
configuration. When the backup starts, it overwrites the existing
Toolkit log file. So if you look in the
ands-vocabs-toolkit.2016-... files from 2016-06-01 onwards, it looks
like the Toolkit hasn't done anything.

Now, specify a separate logging config, and have the backup
script use it.