Add a test for correct persistence of LocalDateTime values
Add a test that persists and fetches a LocalDateTime value that,
considered as a date/time in the Australia/Sydney timezone, does not
exist.

Although we are using LocalDateTime values as dates/times in UTC, we
need to make sure that the values never get interpreted somewhere
along the line as though they were in the Australia/Sydney timezone.

So persist the value LocalDateTime.of(2016, 10, 2, 2, 1, 0) -- which
does not exist as a value in the Australia/Sydney timezone -- and
fetch it back again to see that it comes back correctly.