Show how to increase the level of Hibernate logging
Add comments to logback configuration files to show what must be done
to increase Hibernate logging.  But that isn't the whole story:
because Hibernate uses JBoss Logging, you must add a system property
to tell JBoss logging to use SLF4J. Add a comment to build.xml that
shows how this is done.