Fix errors in definition of related entity in registry schema
There were some errors in the definition of related entities
in the registry schema.
In order to serialize values correctly, the definition
must be of an element, not a complex type.
There were two instances of an underscore where there should
be a hyphen.
Swagger seemed to have been confused by the two different
uses of "related-entity": top-level, and within vocabulary.
So change the one within vocabulary to "related-entity-ref".